+++ src/Classes/EmployerMetadata.php
@@ -21,6 +21,8 @@ class EmployerMetadata extends Data
         public ?Nationality $nationality,
         #[WithCast(EnumCast::class)]
         public ?Industry $industry,
-        public AddressMetadata|Optional $address
+        public AddressMetadata|Optional $address,
+        public ?string $year_established,
+        public ?string $total_number_of_employees,
     ) {}
 }